pip_requirements.txt
absl-py==1.0.0
autopep8==1.6.0
backcall==0.2.0
black==21.12b0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
configparser==5.2.0
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
docker-pycreds==0.4.0
dominate==2.6.0
entrypoints==0.4
fonttools==4.28.5
gitdb==4.0.9
GitPython==3.1.24
google-auth==2.3.3
google-auth-oauthlib==0.4.6
grpcio==1.43.0
idna==3.3
image-slicer==2.1.1
imageio==2.13.5
importlib-metadata==4.10.0
ipykernel==6.13.0
ipython==7.32.0
jedi==0.18.1
jsonpatch==1.32
jsonpointer==2.2
jupyter-client==7.2.2
jupyter-core==4.9.2
kiwisolver==1.3.2
Markdown==3.3.6
matplotlib==3.5.1
matplotlib-inline==0.1.3
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626179032232/work
mkl-service==2.4.0
mypy-extensions==0.4.3
nest-asyncio==1.5.5
networkx==2.6.3
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1634106693478/work
oauthlib==3.1.1
olefile @ file:///Users/ktietz/demo/mc3/conda-bld/olefile_1629805411829/work
opencv-python==4.5.5.62
packaging==21.3
pandas==1.3.5
parso==0.8.3
pathspec==0.9.0
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
platformdirs==2.4.1
promise==2.3
prompt-toolkit==3.0.29
protobuf==3.19.1
psutil==5.9.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
Pygments==2.11.2
pyparsing==3.0.6
python-dateutil==2.8.2
pytorch-fid==0.2.1
pytorch-ignite==0.4.9
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
pyzmq==22.3.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
scikit-image==0.19.1
scipy==1.7.3
seaborn==0.11.2
sentry-sdk==1.5.1
shortuuid==1.0.8
six @ file:///tmp/build/80754af9/six_1623709665295/work
smmap==5.0.0
subprocess32==3.5.4
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
tifffile==2021.11.2
toml==0.10.2
tomli==1.2.3
torch==1.10.0
torch-tb-profiler==0.3.1
torchaudio==0.10.0
torchfile==0.1.0
torchvision==0.11.1
tornado==6.1
traitlets==5.1.1
typed-ast==1.5.1
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1631814937681/work
urllib3==1.26.7
visdom==0.1.8.9
wandb==0.12.9
wcwidth==0.2.5
websocket-client==1.2.3
Werkzeug==2.0.2
yaspin==2.1.0
zipp==3.6.0
